Changes for numpy 2:  - casting changes required for C extensions  - casting numpy float32 scalars to float64 to ensure that arithmetic is done using float64 precision  - remove some conditionals in the C extension that were put in for python 2/3 compatibility

+++ calcos/concurrent.py
@@ -1109,7 +1109,7 @@ class ConcurrentWavecal(object):
             for k in range(k_on, k_off+1):
                 if self.src_counts[k] > cutoff:
                     lamp_is_on = True
-                    t = t0 + k * self.delta_t
+                    t = np.float64(t0) + k * self.delta_t
                     lamp_on.append(t)
                     break
             if not lamp_is_on:
@@ -1117,7 +1117,7 @@ class ConcurrentWavecal(object):
             # search for the actual lamp turn-off time
             for k in range(k_off, k_on-1, -1):
                 if self.src_counts[k] > cutoff:
-                    t = t0 + (k+1) * self.delta_t
+                    t = np.float64(t0) + (k+1) * self.delta_t
                     t = min(t, self.time[-1])
                     lamp_off.append(t)
                     break
@@ -1167,8 +1167,8 @@ class ConcurrentWavecal(object):
             interval (t0, t0+delta_t).
         """
 
-        k_on = (lamp_on_i - t0) / self.delta_t
-        k_off = (lamp_off_i - t0) / self.delta_t
+        k_on = (lamp_on_i - np.float64(t0)) / self.delta_t
+        k_off = (lamp_off_i - np.float64(t0)) / self.delta_t
         k_on = int(math.floor(k_on))
         k_off = int(math.ceil(k_off))
         if k_on >= nbins or k_off < 0:
